move dotprod test into `tests/`
