+++ Cargo.toml
@@ -18,6 +18,7 @@ derive_more = "0.99.17"
 [dev-dependencies]
 criterion = "0.5"
 rand = { version = "0.8.5", features = ["small_rng"] }
+approx = "0.5"
 
 [build-dependencies]
 cfg_aliases = "0.2.0"
